+++ packages/vite/src/node/server/pluginContainer.ts
@@ -57,6 +57,7 @@ import { FSWatcher } from 'chokidar'
 import {
   createDebugger,
   generateCodeFrame,
+  isExternalUrl,
   normalizePath,
   numberToPos,
   prettifyUrl,
@@ -358,6 +359,7 @@ export async function createPluginContainer(
   }
 
   let nestedResolveCall = 0
+  let closed = false
 
   const container: PluginContainer = {
     options: await (async () => {
@@ -459,7 +461,7 @@ export async function createPluginContainer(
       }
 
       if (id) {
-        partial.id = normalizePath(id)
+        partial.id = isExternalUrl(id) ? id : normalizePath(id)
         return partial as PartialResolvedId
       } else {
         return null
@@ -572,6 +574,7 @@ export async function createPluginContainer(
     },
 
     async close() {
+      if (closed) return
       const ctx = new Context()
       await Promise.all(
         plugins.map((p) => p.buildEnd && p.buildEnd.call(ctx as any))
@@ -579,6 +582,7 @@ export async function createPluginContainer(
       await Promise.all(
         plugins.map((p) => p.closeBundle && p.closeBundle.call(ctx as any))
       )
+      closed = true
     }
   }